requirements.txt
# IDShield Dependencies
# Core OCR and Image Processing
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# PDF Processing
PyMuPDF>=1.23.0
PyPDF2>=3.0.0

# Web Interface
streamlit>=1.28.0
streamlit-option-menu>=0.3.0

# Data Processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0

# Machine Learning and AI
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.65.0

# Development and Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0